refactor: extract scroll logic into useScrollManager hook
Move all scroll tracking, MutationObserver, throttling, and
scroll-to-bottom button state into a dedicated hook. Adds debounced
button visibility to prevent flash on conversation load. Throttles
MutationObserver callback to avoid forced reflows during streaming.
